+++ storage/myisam/mi_preload.c
@@ -41,7 +41,7 @@
 int mi_preload(MI_INFO *info, ulonglong key_map, my_bool ignore_leaves)
 {
   uint i;
-  ulong length, block_length= 0;
+  size_t length, block_length= 0;
   uchar *buff= NULL;
   MYISAM_SHARE* share= info->s;
   uint keys= share->state.header.keys;
@@ -84,7 +84,7 @@ int mi_preload(MI_INFO *info, ulonglong key_map, my_bool ignore_leaves)
   {
     /* Read the next block of index file into the preload buffer */
     if ((my_off_t) length > (key_file_length-pos))
-      length= (ulong) (key_file_length-pos);
+      length= (size_t) (key_file_length-pos);
     if (mysql_file_pread(share->kfile, (uchar*) buff, length, pos,
                          MYF(MY_FAE|MY_FNABP)))
       goto err;
